+++ Visual Studio/Tetris_VS/Tetris_VS.sln
@@ -1,0 +1,31 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.5.33530.505
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "Tetris_VS", "Tetris_VS.vcxproj", "{A202B5FD-7A1A-4857-BF6A-4A6346284C55}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|x64 = Debug|x64
+		Debug|x86 = Debug|x86
+		Release|x64 = Release|x64
+		Release|x86 = Release|x86
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{A202B5FD-7A1A-4857-BF6A-4A6346284C55}.Debug|x64.ActiveCfg = Debug|x64
+		{A202B5FD-7A1A-4857-BF6A-4A6346284C55}.Debug|x64.Build.0 = Debug|x64
+		{A202B5FD-7A1A-4857-BF6A-4A6346284C55}.Debug|x86.ActiveCfg = Debug|Win32
+		{A202B5FD-7A1A-4857-BF6A-4A6346284C55}.Debug|x86.Build.0 = Debug|Win32
+		{A202B5FD-7A1A-4857-BF6A-4A6346284C55}.Release|x64.ActiveCfg = Release|x64
+		{A202B5FD-7A1A-4857-BF6A-4A6346284C55}.Release|x64.Build.0 = Release|x64
+		{A202B5FD-7A1A-4857-BF6A-4A6346284C55}.Release|x86.ActiveCfg = Release|Win32
+		{A202B5FD-7A1A-4857-BF6A-4A6346284C55}.Release|x86.Build.0 = Release|Win32
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {B10810B4-0A43-448A-8179-8E2AB829EBD1}
+	EndGlobalSection
+EndGlobal
