+++ Visual Studio/Tetris_VS/x64/Debug/Tetris_VS.log
@@ -1,0 +1,5 @@
+﻿  main.cpp
+  SDL_Plotter.cpp
+  Generating Code...
+MSVCRTD.lib(initializers.obj) : warning LNK4098: defaultlib 'msvcrt.lib' conflicts with use of other libs; use /NODEFAULTLIB:library
+  Tetris_VS.vcxproj -> C:\Users\reese\Documents\CSI1430-Tetris\Visual Studio\Tetris_VS\x64\Debug\Tetris_VS.exe
