+++ Visual Studio/Tetris_VS/x64/Debug/Tetris_VS.tlog/Tetris_VS.lastbuildstate
@@ -1,0 +1,2 @@
+PlatformToolSet=v143:VCToolArchitecture=Native64Bit:VCToolsVersion=14.35.32215:TargetPlatformVersion=10.0.22000.0:VcpkgTriplet=x64-windows:
+Debug|x64|C:\Users\reese\Documents\CSI1430-Tetris\Visual Studio\Tetris_VS\|
