+++ Visual Studio/Tetris_VS/x64/Debug/vcpkg.applocal.log
@@ -1,0 +1,6 @@
+﻿
+C:\Users\reese\Documents\CSI1430-Tetris\Visual Studio\Tetris_VS\x64\Debug\SDL2_mixerd.dll
+C:\Users\reese\Documents\CSI1430-Tetris\Visual Studio\Tetris_VS\x64\Debug\SDL2d.dll
+C:\Users\reese\Documents\CSI1430-Tetris\Visual Studio\Tetris_VS\x64\Debug\vorbisfile.dll
+C:\Users\reese\Documents\CSI1430-Tetris\Visual Studio\Tetris_VS\x64\Debug\vorbis.dll
+C:\Users\reese\Documents\CSI1430-Tetris\Visual Studio\Tetris_VS\x64\Debug\ogg.dll
